requirements.txt
Flask
python-telegram-bot==13.15
stripe